commands refactoring done
