+++ src/commands/backtrace.rs
@@ -13,7 +13,7 @@ impl DebugCommand for BacktraceCommand {
         &["bt"]
     }
 
-    fn execute(&self, args: &[&str], debugger: &mut Debugger) -> Result<()> {
+    fn execute(&self, _args: &[&str], debugger: &mut Debugger) -> Result<()> {
         debugger.backtrace()
     }
 }
