+++ src/commands/sections.rs
@@ -13,7 +13,7 @@ impl DebugCommand for SectionsCommand {
         &["sec"]
     }
 
-    fn execute(&self, args: &[&str], debugger: &mut Debugger) -> Result<()> {
+    fn execute(&self, _args: &[&str], debugger: &mut Debugger) -> Result<()> {
         debugger.print_sections()
     }
 }
